Moved to migrating ruleset
