+++ pmd-java/src/main/resources/rulesets/java/migrating.xml
@@ -473,4 +473,29 @@ public class MyTest {
       </example>
     </rule>
 
+    <rule name="ForLoopCanBeForeach"
+          language="java"
+          since="6.0"
+          message="This 'for' loop can be replaced by a 'foreach' loop"
+          class="net.sourceforge.pmd.lang.java.rule.migrating.ForLoopCanBeForeachRule"
+          externalInfoUrl="${pmd.website.baseurl}/rules/java/migrating.html#ForLoopCanBeForeach">
+        <description>
+            TODO
+        </description>
+        <priority>3</priority>
+        <example><![CDATA[
+public class MyClass {
+  void loop(List<String> l) {
+    for (int i = 0; i < l.size(); i++) { // pre Java 1.5
+      System.out.println(l.get(i));
+    }
+
+    for (String s : l) {        // post Java 1.5
+      System.out.println(s);
+    }
+  }
+}
+        ]]></example>
+    </rule>
+
 </ruleset>
