+++ pmd-java/src/test/resources/net/sourceforge/pmd/lang/java/rule/migrating/xml/ForLoopCanBeForeach.xml
@@ -225,8 +225,4 @@
         ]]></code>
     </test-code>
 
-
-
-
-
 </test-data>
